+++ src/providers/twitch/TwitchIrcServer.cpp
@@ -28,6 +28,7 @@ TwitchIrcServer::TwitchIrcServer()
     : whispersChannel(new Channel("/whispers", Channel::Type::TwitchWhispers))
     , mentionsChannel(new Channel("/mentions", Channel::Type::TwitchMentions))
     , watchingChannel(Channel::getEmpty(), Channel::Type::TwitchWatching)
+    , liveChannel(new Channel("/live", Channel::Type::TwitchLive))
 {
     this->initializeIrc();
 
@@ -257,6 +258,11 @@ std::shared_ptr<Channel> TwitchIrcServer::getCustomChannel(
         return this->mentionsChannel;
     }
 
+    if (channelName == "/live")
+    {
+        return this->liveChannel;
+    }
+
     if (channelName == "$$$")
     {
         static auto channel =
@@ -289,6 +295,7 @@ void TwitchIrcServer::forEachChannelAndSpecialChannels(
 
     func(this->whispersChannel);
     func(this->mentionsChannel);
+    func(this->liveChannel);
 }
 
 std::shared_ptr<Channel> TwitchIrcServer::getChannelOrEmptyByID(