+++ src/widgets/helper/ChannelView.cpp
@@ -1,6 +1,7 @@
 #include "ChannelView.hpp"
 
 #include <QClipboard>
+#include <QDate>
 #include <QDebug>
 #include <QDesktopServices>
 #include <QGraphicsBlurEffect>
@@ -22,6 +23,7 @@
 #include "messages/Emote.hpp"
 #include "messages/LimitedQueueSnapshot.hpp"
 #include "messages/Message.hpp"
+#include "messages/MessageBuilder.hpp"
 #include "messages/MessageElement.hpp"
 #include "messages/layouts/MessageLayout.hpp"
 #include "messages/layouts/MessageLayoutElement.hpp"
@@ -602,6 +604,15 @@ void ChannelView::setChannel(ChannelPtr underlyingChannel)
                    boost::optional<MessageFlags> overridingFlags) {
                 if (this->shouldIncludeMessage(message))
                 {
+                    if (this->channel_->lastDate_ != QDate::currentDate())
+                    {
+                        this->channel_->lastDate_ = QDate::currentDate();
+                        auto msg =
+                            makeSystemMessage(QDate::currentDate().toString(
+                                                  Qt::SystemLocaleLongDate),
+                                              QTime(0, 0));
+                        this->channel_->addMessage(msg);
+                    }
                     // When the message was received in the underlyingChannel,
                     // logging will be handled. Prevent duplications.
                     if (overridingFlags)
@@ -994,7 +1005,8 @@ MessageElementFlags ChannelView::getFlags() const
         {
             flags.set(MessageElementFlag::ModeratorTools);
         }
-        if (this->underlyingChannel_ == app->twitch.server->mentionsChannel)
+        if (this->underlyingChannel_ == app->twitch.server->mentionsChannel ||
+            this->underlyingChannel_ == app->twitch.server->liveChannel)
         {
             flags.set(MessageElementFlag::ChannelName);
             flags.unset(MessageElementFlag::ChannelPointReward);